Made the main window an object

+++ CsvReader.h
@@ -9,17 +9,17 @@ class CsvReader
     public:
     CsvReader(QFile* in);
     CsvReader(QFile* in, int PLACEHOLDER);
-    QString get(int row, int column);
+    QString get(size_t row, size_t column);
     size_t getNumRows();
     void saveFile();
     void saveFileNoClose();
-    void set(int row, int column, std::string val);
+    void set(size_t row, size_t column, std::string val);
     void addRow();
     void addRow(std::vector<std::string> init);
-    std::vector<std::string> getRow(int row);
-    void addToRow(int row, std::string val);
-    void deleteRow(int row);
-    void deleteRows(int begin, int end);
+    std::vector<std::string> getRow(size_t row);
+    void addToRow(size_t row, std::string val);
+    void deleteRow(size_t row);
+    void deleteRows(size_t begin, size_t end);
     void deleteAll();
     
     private:
